realtime-interview-avatar/requirements.txt
# ============================================================================
# Core Dependencies
# ============================================================================
python-dotenv==1.0.0         # 환경변수 관리 (.env 파일 로딩)
pydantic==2.5.0              # 데이터 검증 및 설정 관리
pydantic-settings==2.1.0     # Pydantic 기반 설정 관리

# ============================================================================
# Pipecat - Real-time Pipeline Framework
# ============================================================================
pipecat-ai[daily,deepgram,elevenlabs,openai,silero]==0.0.43
# - 실시간 오디오/비디오 처리 파이프라인 프레임워크
# - daily: Daily.co WebRTC 통합
# - deepgram: Deepgram STT 통합
# - elevenlabs: ElevenLabs TTS 통합
# - openai: OpenAI API 통합
# - silero: Silero VAD (Voice Activity Detection) 통합

# ============================================================================
# LLM Integration
# ============================================================================
openai>=1.0.0                # OpenAI API (GPT-4 등)
anthropic>=0.18.0            # Anthropic API (Claude 등)

# ============================================================================
# Alternative TTS (한국어 TTS 대안)
# ============================================================================
edge-tts>=6.1.0              # Microsoft Edge TTS (무료, 한국어 지원)

# ============================================================================
# Deep Learning Framework (MuseTalk 아바타)
# ============================================================================
torch==2.0.1                 # PyTorch 딥러닝 프레임워크
torchvision==0.15.2          # 컴퓨터 비전 유틸리티
torchaudio==2.0.2            # 오디오 처리 유틸리티
diffusers==0.21.0            # Hugging Face Diffusion 모델 (Stable Diffusion 등)
transformers==4.36.0         # Hugging Face Transformers (언어/비전 모델)
accelerate==0.25.0           # 분산 학습 및 추론 가속화

# ============================================================================
# Computer Vision & Face Processing (아바타 생성)
# ============================================================================
opencv-python==4.8.1.78      # 이미지/비디오 처리
mediapipe==0.10.8            # 얼굴/손/포즈 랜드마크 검출
insightface==0.7.3           # 얼굴 인식 및 분석
onnxruntime-gpu==1.16.3      # ONNX 모델 GPU 추론

# ============================================================================
# Audio Processing
# ============================================================================
librosa==0.10.1              # 오디오 분석 및 특징 추출
soundfile==0.12.1            # 오디오 파일 읽기/쓰기

# ============================================================================
# Video Processing
# ============================================================================
ffmpeg-python==0.2.0         # FFmpeg Python 바인딩 (비디오 인코딩/디코딩)
av==11.0.0                   # PyAV - 비디오/오디오 컨테이너 처리

# ============================================================================
# Face Enhancement (선택사항)
# ============================================================================
gfpgan==1.3.8                # 얼굴 복원 및 향상 (GAN 기반)
realesrgan==0.3.0            # 이미지 초해상도 복원 (Real-ESRGAN)

# ============================================================================
# WebRTC & Networking
# ============================================================================
aiortc==1.6.0                # WebRTC 구현 (브라우저와 실시간 통신)
aiohttp==3.9.1               # 비동기 HTTP 클라이언트/서버

# ============================================================================
# Web Framework (API 서버)
# ============================================================================
fastapi>=0.104.0             # 현대적인 웹 API 프레임워크
uvicorn>=0.24.0              # ASGI 서버 (FastAPI 실행)
websockets>=12.0             # WebSocket 프로토콜 구현

# ============================================================================
# Utilities
# ============================================================================
loguru==0.7.2                # 간편한 로깅 라이브러리
tqdm==4.66.1                 # 프로그레스 바
numpy==1.24.3                # 수치 계산 라이브러리
pillow==10.1.0               # 이미지 처리 라이브러리 (PIL)
pyyaml>=6.0.0                # YAML 파일 파싱
aiofiles>=23.0.0             # 비동기 파일 I/O
requests>=2.31.0             # HTTP 요청 라이브러리
httpx>=0.25.0                # 비동기 HTTP 클라이언트

# ============================================================================
# Development & Testing
# ============================================================================
pytest>=7.4.0                # 테스트 프레임워크
pytest-asyncio>=0.21.0       # 비동기 테스트 지원
black>=23.0.0                # 코드 포맷터
flake8>=6.0.0                # 코드 린터
mypy>=1.5.0                  # 정적 타입 체커
